DAY 5 + 6

+++ src/main/java/com/social/network/Services/PostService.java
@@ -1,8 +1,12 @@
 package com.social.network.Services;
 
 import com.social.network.Dao.GenericDao;
+import com.social.network.Dto.PostDTO;
+import com.social.network.Facade.PostFacade;
 import com.social.network.Model.Post;
 import com.social.network.Model.User;
+import com.social.network.Services.Interfaces.IPostService;
+import com.social.network.exceptions.NoPermission;
 import com.social.network.exceptions.PostNotFoundException;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -11,11 +15,13 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.EntityManager;
-import javax.persistence.criteria.CriteriaBuilder;
-import javax.persistence.criteria.CriteriaQuery;
-import javax.persistence.criteria.Predicate;
-import javax.persistence.criteria.Root;
+import javax.persistence.criteria.*;
 import java.security.Principal;
+import java.util.Comparator;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 @Service
 @Transactional
@@ -25,43 +31,81 @@ public class PostService implements IPostService {
 
     private GenericDao<Post> postGenericDao;
 
+    private PostFacade postFacade;
+
     @Autowired
     public PostService(GenericDao<Post> postGenericDao) {
         this.postGenericDao = postGenericDao;
         log = LogManager.getLogger(PostService.class);
     }
 
+    @Autowired
+    public void setPostFacade(PostFacade postFacade) {
+        this.postFacade = postFacade;
+    }
+
     @Override
-    public Post createNewPostFromUser(Post post, Principal principal) {
+    public PostDTO createNewPostFromUser(Post post, Principal principal) {
         User timeUser = findByPrincipal(principal.getName());
-        timeUser.getPosts().add(post);
+        post.setUser(timeUser);
         postGenericDao.add(post);
         log.info("User " + timeUser.getUsername() + "create new post");
-        return post;
+        PostDTO postDTO = postFacade.postToPostDTO(post);
+        return postDTO;
     }
 
     @Override
-    public boolean deleatePost(int postID) {
+    public boolean deleatePost(int postID,Principal principal) {
+        User user = findByPrincipal(principal.getName());
         Post post = postGenericDao.find(postID);
         if (post != null) {
-            postGenericDao.delete(postID);
-            return true;
+            if(post.getUser().equals(user)){
+                postGenericDao.delete(postID);
+                return true;
+            } else {
+                throw new NoPermission("You cant do it");
+            }
         } else {
             throw new PostNotFoundException("Did not found post with such id: " + postID);
         }
     }
 
     @Override
-    public Post findPostByID(int postID) {
+    public PostDTO findPostByID(int postID) {
         Post post = postGenericDao.find(postID);
         if (post != null) {
-            return post;
+            return postFacade.postToPostDTO(post);
         } else {
             throw new PostNotFoundException("Did not found post with such id: " + postID);
         }
     }
 
-     private User findByPrincipal(String name) {
+    @Override
+    public List<PostDTO> openNews(Principal principal){
+        User user = findByPrincipal(principal.getName());
+        List<User> users = getFriendsToOpenNews(user);
+        Set<Post> posts = new HashSet<>();
+        for (User us : users) {
+            posts.addAll(us.getPosts());
+        }
+        return posts.stream().map(postFacade::postToPostDTO).
+                sorted(Comparator.comparing(PostDTO::getCreatedDate)).collect(Collectors.toList());
+    }
+
+    private List<User> getFriendsToOpenNews(User userTest){
+        EntityManager em = postGenericDao.getEntityManager();
+        CriteriaBuilder cb = em.getCriteriaBuilder();
+        CriteriaQuery<User> query = cb.createQuery(User.class);
+        Root<User> user = query.from(User.class);
+        Join<User, User> friendsJoin = user.join("friends");
+        Predicate userPredicate = cb.equal(friendsJoin.get("id"), userTest.getId());
+        query.select(user).where(userPredicate);
+        List<User> users = em.createQuery(query).getResultList();
+        users.removeIf(us -> !userTest.getFriends().contains(us));
+        return users;
+    }
+
+    private User findByPrincipal(String name) {
         User timeUser = null;
         try {
             EntityManager em = postGenericDao.getEntityManager();
